+++ drizzle/0009_cultured_owl.sql
@@ -1,0 +1,3 @@
+CREATE TYPE "public"."invitation_status" AS ENUM('pending', 'accepted', 'rejected');--> statement-breakpoint
+ALTER TABLE "invitation" ADD COLUMN "status" "invitation_status" DEFAULT 'pending' NOT NULL;--> statement-breakpoint
+ALTER TABLE "invitation" DROP COLUMN "accepted";
