+++ package.json
@@ -30,6 +30,7 @@
     "@radix-ui/react-tabs": "^1.1.3",
     "@react-email/components": "0.0.33",
     "@tailwindcss/postcss": "^4.0.7",
+    "@tanstack/react-query": "^5.71.1",
     "better-auth": "^1.1.18",
     "better-auth-harmony": "^1.2.3",
     "class-variance-authority": "^0.7.1",
