+++ src/app/(auth)/reset-password/page.tsx
@@ -10,7 +10,8 @@ import { z } from "zod"
 
 import { resetPassword } from "@/lib/auth/auth-client"
 import { cn } from "@/lib/utils"
-import { Button, buttonVariants } from "@/components/ui/button"
+import { Button } from "@/components/ui/button"
+import { buttonVariants } from "@/components/ui/button-variants"
 import {
   Card,
   CardContent,
