+++ src/app/(auth-protected)/events/create/page.tsx
@@ -13,11 +13,11 @@ export default function CreateEventPage() {
   const handleSubmit = async (data: z.infer<typeof eventDetailsSchema>) => {
     // Create the event and navigate to the event management page
     const eventId = await createEvent(data)
-    router.push(`/events/${eventId}`)
+    router.push(`/dashboard/events/${eventId}`)
   }
 
   return (
-    <main className="max-w-3xl mx-auto p-6">
+    <main className="max-w-3xl mx-auto p-6 w-full">
       <div className="flex items-center justify-between mb-6">
         <h1 className="text-2xl font-bold">Create New Event</h1>
       </div>
